medicine_app/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Werkzeug==2.3.7
mysqlclient==2.2.0
pymysql==1.1.0
requests==2.31.0
python-dotenv==1.0.0